Add support for 64-bit indexes.

+++ pyfr/plugins/base.py
@@ -61,7 +61,7 @@ def region_data(cfg, cfgsect, mesh, rallocs):
         if not comm.reduce(bool(eset), op=mpi.LOR, root=root) and rank == root:
             raise ValueError(f'Empty region {region}')
 
-        return {etype: np.unique(eidxs).astype(np.int32)
+        return {etype: np.unique(eidxs)
                 for etype, eidxs in sorted(eset.items())}
